Sucessfully building voc labels
