+++ .gitignore
@@ -18,6 +18,7 @@ notes.txt
 *tif
 *xml
 notebooks/Scratch.ipynb
+Untitled.ipynb
 
 # temp files
 *swp
